+++ benchmarking/benchmarking_dpo.py
@@ -35,8 +35,8 @@ from agilerl.algorithms.dpo import DPO
 from agilerl.hpo.mutation import Mutations
 from agilerl.hpo.tournament import TournamentSelection
 from agilerl.training.train_llm import finetune_llm_preference
+from agilerl.wrappers.llm_envs import PreferenceGym
 from agilerl.utils.llm_utils import (
-    PreferenceGym,
     compare_responses,
     sample_eval_prompts,
 )
