+++ docs/api/wrappers/index.rst
@@ -5,6 +5,7 @@ Wrappers
    :maxdepth: 1
 
    agent
+   llm_envs
    make_evolvable
    learning
    pettingzoo
